requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
qdrant-client>=1.7.0
sentence-transformers>=2.2.0
youdotcom>=1.4.0
httpx>=0.26.0
python-dotenv>=1.0.0
